refactor(mcp): split the two /mcp three.js scene modules
`PlaygroundViewer.tsx` (926 lines) and `HeroScene.tsx` (891) were both far
past the ~400 non-generated line guideline in AGENTS.md. Both had the same
shape: a small React component on top, a large imperative `createScene`
factory underneath, meeting only at a `SceneHandle` interface.

Extracting `createScene` alone does not reach compliance — it leaves the
playground scene at ~561 lines — so each half is decomposed along the
concerns the scene already had.

Playground (926 -> 307 + 92 + 294 + 213 + 138 + 139):

  PlaygroundViewer.tsx          React half + `loadPlaygroundGeometry`
  playground-viewer-types.ts    ViewerController / SceneHandle interfaces
  playground-scene.ts           renderer, picking, handle assembly
  playground-scene-registry.ts  per-entity meshes + lookup maps
  playground-scene-ops.ts       colorize / isolate / hide / show / reset
  playground-scene-view.ts      camera framing + section planes

Hero (891 -> 145 + 213 + 368 + 313):

  HeroScene.tsx                 React half + the HERO_STEPS data
  hero-scene.ts                 renderer, lights, animation loop
  hero-scene-building.ts        building geometry, section prop, BCF pin
  hero-scene-steps.ts           the twelve-step story arc

The move is mechanical. Function bodies are unchanged; the only edits are
the ones the extraction forces: closed-over state passed explicitly
(`EntityRegistry`, `SectionState`, `SceneView`, `HeroAnimationState`),
constants exported where a sibling module now reads them, and imports
adjusted. No renames, no reordered arguments, no tightened types.

Nothing here is covered by a test: happy-dom has no WebGL, so neither
`createScene` executes in any suite. Verification is typecheck plus reading
the diff, not tests.
