+++ apps/viewer/AGENTS.md
@@ -10,5 +10,5 @@ Supplements the root [AGENTS.md](../../AGENTS.md); the shared contract (canonica
 - **CSV/list exports neutralize formula injection in two places.** `lib/lists/export/csv.ts` `esc()` (also strips a leading BOM first) and `lib/search/result-export.ts` `escapeCsvCell()` both prefix a leading `= + - @ TAB CR` with an apostrophe (CWE-1236). Any new CSV/list export of attacker-controllable IFC values must reuse one of these, not add a third.
 - **Save files through `lib/export/download.ts`** (`downloadBlob` / `downloadFile` / `downloadDataUrl`); never hand-roll an `<a download>` blob dance. Pass the stem through `sanitizeFilename` (preserves case and dots so `DRAWINGS` and `000.000` survive, #1299), then append the extension yourself. Every download fires `emitFileDownloaded(filename)`, which task-gated tours observe; keep that seam.
 - **New top-right viewport overlays anchor at `top-32 right-4` or lower.** The ViewCube sits at `top-6 right-6` (`components/viewer/ViewportOverlays.tsx`); the existing top-right panels (SunSky, level-display selector) clear it deliberately. Never cover the ViewCube corner.
-- **three.js / Cesium scenes need double-sided materials.** The MCP `PlaygroundViewer` / `HeroScene` and the Cesium GLB export must set `side: THREE.DoubleSide` / `doubleSided: true` on every material; IFC winding is not reliably outward, so back-face culling shows see-through walls. This is separate from the main WebGPU renderer (which lives in `packages/*`, not here).
+- **three.js / Cesium scenes need double-sided materials.** The MCP scenes (`playground-scene-registry.ts` builds the playground's per-entity materials, `hero-scene-building.ts` the hero's) and the Cesium GLB export must set `side: THREE.DoubleSide` / `doubleSided: true` on every material; IFC winding is not reliably outward, so back-face culling shows see-through walls. This is separate from the main WebGPU renderer (which lives in `packages/*`, not here).
 - **Never construct a GPU context straight from a mount effect.** `new THREE.WebGLRenderer` and `new maplibregl.Map` both throw when the device refuses a context, and a throw inside `useEffect` unwinds to the nearest error boundary — which on `/mcp` is the route-level `ChunkErrorBoundary`, so one dead canvas replaced a whole page of GPU-free content with an unrecoverable "Reload" card (#1914, #2401). Mount three.js scenes through `components/mcp/useThreeScene.ts` and maps through `LocationMap.tsx`'s `degradeMap`; both ride the one session-latched gate in `lib/webgl-capability.ts` (probe once, latch the verdict, report ONE handled exception per session). Don't add a third probe or a second latch — a device that refuses a context refuses it for the whole session, and re-probing burns one of the ~16 context slots the page gets. A guard must still rethrow anything that is not the library's own context-refusal message, or a real bug in scene code gets relabelled "your device cannot do 3D" and disappears from error tracking.
